Certifier

Run the Agda certifier on the simplification trace, if requested

prettyCertifierError :: CertifierError -> String

data CertifierError

Constructors

InvalidCertificate CertDir

InvalidCompilerOutput

ValidationError CertName

data CertifierOutput

Constructors

BasicOutput
Print minimal basic info, such as "passed" or "failed"

ReportOutput FilePath
Produce a detailed human readable report

ProjectOutput CertDir
Produce an Agda project that can be type checked